+++ api/controllers/state/break.js
@@ -43,7 +43,7 @@ module.exports = {
       })()
 
       // If this break was triggered because of a technical problem, play a technical problem liner
-      if (inputs.problem) { await sails.helpers.songs.queue(sails.config.custom.subcats.technicalIssues, 'top', 1) }
+      if (inputs.problem) { await sails.helpers.songs.queue(sails.config.custom.subcats.technicalIssues, 'top', 1, 'noRules') }
 
       // halftime break? Play a station ID and then begin halftime music
       if (inputs.halftime) {